==> Useful Methods/2_Remove_Single_Criticals.ipynb <==
Step 2: Removing single critical alarms
Some critical alarms occur without any other alarm type beforehand. These lines need to get removed

import pandas as pd

def delete_single_criticals(df):
    previous_row_severity = df.iloc[0][2]
    rows_to_remove = []

    for index, row in df.iterrows():
        if row.iloc[2] == 'CRITIQUE':
            if previous_row_severity == 'CRITIQUE':
                rows_to_remove.append(index)
        previous_row_severity = row.iloc[2]

    print("Deleting rows...")
    df2 = df.drop(rows_to_remove, inplace=False)
    return df2

for i in range (1,26):
    print(str(i) + " out of 25")
    df = pd.read_csv("..Data/1 Incomplete Sequences Removed/" + str(i) + "_incomplete_removed.csv", delimiter=',')
    df = delete_single_criticals(df)
    df.to_csv(str(i) + "_without_single_criticals.csv", index=False)
