==> Useful Methods/3_Extract_Last_20.ipynb <==
Step 3: Transformation into sequences
Within the project, only the last 20 alarm logs before a critical alarm are taken into account. This notebook extracts the last 20 alarms of each log sequence and transforms it into a more simple representation. In the end, each sequence consists of 20 alarm IDs and finishes with the ID of the critical alarm type to which the sequence has led. If sequences are shorter than 20, they get filled up with 0 values to ensure that each sequence is of the same length.

def prepare_data():
    print("Preparing data...")
    prepared_x = []
    prepared_y = []
    current_sequence = []
    counter = 1

    for index, row in df.iterrows():
        # Each sequence finishes with a critical alarm. We only take into account the last 20 alarms before a critical alarm
        if row.iloc[2] == "CRITIQUE":
            if len(current_sequence) > 20:
                prepared_x.append(current_sequence[-20:])
            else:
                prepared_x.append(current_sequence)
            prepared_y.append(int(row.iloc[1]))
            current_sequence = []
        else:
            current_sequence.append(int(row.iloc[1]))
        counter = counter+1
    return prepared_x, prepared_y

# To ensure that each sequence has the same length, sequences with a length of less than 20 are filled with 0 values
def fill_sequences_with_zero():
    print("Filling with 0...")
    new_X_temp = []
    for item in X_temp:
        for i in range(len(item), 20):
            item.append(0)
        new_X_temp.append(item)

    return new_X_temp

def add_to_X(x_to_add):
    for item in x_to_add:
        X.append(item)

def add_to_y(y_to_add):
    for item in y_to_add:
        y.append(item)

import pandas as pd

X = []
y = []

for i in range(1,26):
    print(str(i) + " out of 25")
    df = pd.read_csv("../Data/2 Clean without single criticals/" + str(i) + "_without_single_criticals.csv", header=None)
    X_temp, y_temp = prepare_data()
    X_temp = fill_sequences_with_zero()
    add_to_X(X_temp)
    add_to_y(y_temp)
    df2 = pd.DataFrame(X)
    df2['y'] = y
    print("Saving new file for " + str(i))
    df2.to_csv(str(i) + "_last20_sequences.csv", index=False)
    X = []
    y = []
