==> Useful Methods/4_Separate_Error_Types.ipynb <==
Step 4: Separate data by alarm type
The available samples are sorted by alarm type and saved in separate .csv files

def separate_error_types():
    for index,row in df.iterrows():
        y = row.iloc[100]
        if y == 7:
            samples_7.append(row)
        elif y == 15:
            samples_15.append(row)
        elif y == 16:
            samples_16.append(row)
        elif y == 21:
            samples_21.append(row)
        elif y == 33:
            samples_33.append(row)
        elif y == 34:
            samples_34.append(row)
        elif y == 39:
            samples_39.append(row)
        elif y == 56:
            samples_56.append(row)
        elif y == 68:
            samples_68.append(row)
        elif y == 95:
            samples_95.append(row)
        elif y == 1000:
            samples_1000.append(row)
        elif y == 1001:
            samples_1001.append(row)
        else:
            continue

import pandas as pd

samples_7 = []
samples_15 = []
samples_16 = []
samples_21 = []
samples_33 = []
samples_34 = []
samples_39 = []
samples_56 = []
samples_68 = []
samples_95 = []
samples_1000 = []
samples_1001 = []

total_length = 0
current_length = 0

for i in range(1,26):
    print(str(i) + " out of 25")
    df = pd.read_csv("../Data/3 Transformed into Sequences/" + str(i) + "_sequences100.csv")
    current_length = len(df)
    print("Length: ", current_length)
    total_length = total_length + current_length
    separate_error_types()
    current_length = 0

print("Total length: ", total_length)
df7 = pd.DataFrame(samples_7)
df15 = pd.DataFrame(samples_15)
df16 = pd.DataFrame(samples_16)
df21 = pd.DataFrame(samples_21)
df33 = pd.DataFrame(samples_33)
df34 = pd.DataFrame(samples_34)
df39 = pd.DataFrame(samples_39)
df56 = pd.DataFrame(samples_56)
df68 = pd.DataFrame(samples_68)
df95 = pd.DataFrame(samples_95)
df1000 = pd.DataFrame(samples_1000)
df1001 = pd.DataFrame(samples_1001)

df7.to_csv("7.csv", index=None)
df15.to_csv("15.csv", index=None)
df16.to_csv("16.csv", index=None)
df21.to_csv("21.csv", index=None)
df33.to_csv("33.csv", index=None)
df34.to_csv("34.csv", index=None)
df39.to_csv("39.csv", index=None)
df56.to_csv("56.csv", index=None)
df68.to_csv("68.csv", index=None)
df95.to_csv("95.csv", index=None)
df1000.to_csv("1000.csv", index=None)
df1001.to_csv("1001.csv", index=None)
