==> Useful Methods/5_Transform_to_countbased.ipynb <==
Step 5 (optional): Transform sequences into count-based encoding

import pandas as pd

dictionary_frame = {1:0,2:0,3:0,4:0,5:0,9:0,10:0,11:0,13:0,14:0,17:0,19:0,20:0,22:0,23:0,24:0,
                    25:0,26:0,27:0,28:0,29:0,30:0,31:0,32:0,35:0,36:0,37:0,38:0,40:0,41:0,42:0,43:0,44:0,45:0,
                    46:0,47:0,48:0,49:0,50:0,52:0,53:0,57:0,58:0,59:0,60:0,62:0,63:0,64:0,65:0,66:0,
                    67:0,69:0,70:0,71:0,79:0,81:0,83:0,84:0,85:0,86:0,87:0,88:0,89:0,90:0,91:0,92:0,93:0,94:0,96:0,
                    98:0,99:0,100:0,101:0,102:0,103:0,104:0,107:0}

error_list = [7,15,16,21,33,56,68,95,1000,1001]

def transform_to_countbased(current_x):
    dictionary_current = dictionary_frame
    dictionary_current = dict.fromkeys(dictionary_current, 0)
    list_data = []
    for item in current_x:
        for sample in item:
            if not sample == 0:
                dictionary_current[sample] += 1
        list_data.append(dictionary_current)
        dictionary_current = dict.fromkeys(dictionary_current, 0)
    return list_data


# Adapt link depending on which dataset you want to transform. All, Unique Samples or Random Samples
for i in error_list:
    y = []
    df = pd.read_csv("../Data/Train/Unique Samples/Sequences/" + str(i) + ".csv")
    y = df['y'].values
    x = df.drop(columns=['y']).values
    data = transform_to_countbased(x)
    df_new = pd.DataFrame(data)
    df_new['y'] = y
    df_new.to_csv(str(i) + '_countbased.csv', index=False)
